requirements.txt
google-api-python-client
google-auth
requests
python-dotenv
google-auth-oauthlib
google-auth-httplib2